Implemented functionality to show all user images
backend API is still missing for this to be tested. Need to be integrated once backend is ready.
The button in FinalPage navigates here. There was no need to trigger the get images for a user api from the button call.
Once this view is navigated to, it automatically calls the API from useEffect.